Put PNG files in extra-doc-files
Currently, `lens` uses custom `Setup` logic to embed its PNG files
into Haddocks, but nowadays there is a `Cabal` option for this very
thing: `extra-doc-files`. Using this lets us simplify the custom
`Setup` logic considerably (in hopes of eventually removing it one
day, once `cabal doctest` is a thing).

The downside to this is that `lens` will need `Cabal-1.18` as a
minimum to build. Is this worth it?